requirements.txt
django
django-bootstrap3
django-registration-redux
ipython
django-extensions
djangorestframework
